[Solve] 8주차 폴더구조 변경
