+++ src/main/java/week8/baek/june25/B10988.java
@@ -1,17 +1,20 @@
-package week8.baek;
+package week8.baek.june25;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class B11721 {
+public class B10988 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         String input = br.readLine();
 
-        for (int i = 0; i < input.length(); i+=10) {
-            if(i+10 < input.length()) System.out.println(input.substring(i, i + 10));
-            else System.out.println(input.substring(i));
+        String reverse = "";
+        for (int i = input.length() -1; i >= 0; i--) {
+            reverse += input.charAt(i);
         }
+
+        if(reverse.equals(input)) System.out.println(1);
+        else System.out.println(0);
     }
 }
