+++ src/main/java/week8/baek/june25/B11721.java
@@ -1,16 +1,17 @@
-package week8.baek;
+package week8.baek.june25;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class B1598 {
+public class B11721 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         String input = br.readLine();
-        int n = Integer.parseInt(input.split(" ")[0]) - 1;
-        int m = Integer.parseInt(input.split(" ")[1]) - 1;
 
-        System.out.println(Math.abs(m%4 - n%4) + Math.abs(n/4-m/4));
+        for (int i = 0; i < input.length(); i+=10) {
+            if(i+10 < input.length()) System.out.println(input.substring(i, i + 10));
+            else System.out.println(input.substring(i));
+        }
     }
 }
