+++ src/main/java/week8/baek/june25/B1598.java
@@ -1,20 +1,16 @@
-package week8.baek;
+package week8.baek.june25;
 
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
 
-public class B10988 {
+public class B1598 {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         String input = br.readLine();
+        int n = Integer.parseInt(input.split(" ")[0]) - 1;
+        int m = Integer.parseInt(input.split(" ")[1]) - 1;
 
-        String reverse = "";
-        for (int i = input.length() -1; i >= 0; i--) {
-            reverse += input.charAt(i);
-        }
-
-        if(reverse.equals(input)) System.out.println(1);
-        else System.out.println(0);
+        System.out.println(Math.abs(m%4 - n%4) + Math.abs(n/4-m/4));
     }
 }
