+++ src/main/java/week8/baek/june25/B2588.java
@@ -1,4 +1,4 @@
-package week8.baek;
+package week8.baek.june25;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -21,6 +21,7 @@ public class B2588 {
 
             total += tmp;
         }
+
         System.out.println(total);
 
     }
